// src/app/shared/interfaces/post-atendimentos.model.ts
import { AcessoriosAtendimentoMV } from './acessorios-atendimento.model';
import { ChecklistDocumentosChavesMV } from './checklist-documentos-chaves.model';
import { FornecedorAtendimentoMV } from './fornecedor-atendimento.model';

export interface PostAtendimentosMV {
  veiculoId?: number;
  clienteId?: number;
  usuarioId?: number;
  km?: number;
  municipioId?: number;
  motorista?: string;
  telefone?: string;
  email?: string;
  descricaoOcorridoGeral?: string;
  checklist?: ChecklistDocumentosChavesMV[];
  atendimentosFornecedor?: FornecedorAtendimentoMV[];
  atendimentosAcessorios?: AcessoriosAtendimentoMV[];
  tipo?: string;
  agendamentoPrevio?: string;
  direcionadoOficina?: string;
  registrarAtualizacaoKm?: boolean;
  atendimentoId?: number;
  enderecoLocalCliente?: string;
  celular?: string;
  formaContatoId?: number;
  cpfCondutor?: string;
  dataParada1?: number;
  dataParada2?: number;
  dataParada3?: number;
  horaParada1?: number;
  horaParada2?: number;
  horaParada3?: number;
  sinistro?: boolean;
  manutencaoPreventiva?: boolean;
  manutencaoCorretiva?: boolean;
  planoManutencaoId?: number;
  socorroMecanico?: boolean;
  fornecedorId?: number;
  logradouroSinistro?: string;
  bairroSinistro?: string;
  cidadeSinistro?: string;
  municipioSinistroId?: number;
  estadoSinistro?: string;
  cepSinistro?: string;
  boletimOcorrenciaSinistro?: string;
  dataSinistro?: number;
  horaSinistro?: string;
  houveVitima?: boolean;
  houveVitimaFatal?: boolean;
  descricaoOcorridoSinistro?: string;
  motivoTipo?: number;
  nomeTerceiro?: string;
  telefoneTerceiro?: string;
  confirmacaoDocumentacaoNp?: string;
  confirmacaoAVS?: string;
  confirmacaoBO?: string;
  confirmacaoDUT?: string;
  confirmacaoCNH?: string;
  confirmacaoFoto?: string;
  semAgendamentoParada?: number;
  sinistroAnteriorId?: number;
  segmentoNegocioId?: number;
  atendimentoLoja?: boolean;
  siglaLojaUnidas?: string;
}